Modify struct JSONParser not to use RegEXP
It's inefficient to use RegEXP not only the length of code, but also the increasing number of RegExp for each type.
Therefore, move String extension to InputView only to use RegEXP for json pattern checking of user input.
Instead, add searchJsonData() to get elements of String in json standard.
Modify parse() func to use optional binding when casting type.